Better skipping control
git-svn-id: https://appletv.nanopi.net/svn/trunk@186 7e1782f7-7b3f-45d9-b215-e5100ed717cc

+++ SapphireAudioPlayer.h
@@ -12,7 +12,7 @@
 	QTMovie		*movie;
 	int			state;
 	NSTimer		*updateTimer;
-	int			skipSpeed;
+	float		skipSpeed;
 	NSTimer		*skipTimer;
 }
